+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_multiapi_async.py
@@ -37,11 +37,11 @@ class TestMultiapi(AsyncFormRecognizerTest):
         assert "v2.0" in client._client._client._base_url
 
     @GlobalFormRecognizerAccountPreparer()
-    @FormRecognizerClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_1_PREVIEW_1})
+    @FormRecognizerClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_1_PREVIEW})
     def test_v2_1_preview_1_form_recognizer_client(self, client):
         assert "v2.1-preview.1" in client._client._client._base_url
 
     @GlobalFormRecognizerAccountPreparer()
-    @FormTrainingClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_1_PREVIEW_1})
+    @FormTrainingClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_1_PREVIEW})
     def test_v2_1_preview_1_form_training_client(self, client):
         assert "v2.1-preview.1" in client._client._client._base_url