+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/certificate.py
@@ -41,7 +41,7 @@ class CertificateCredential(CertificateCredentialBase, AsyncContextManager):
     async def get_token(self, *scopes: str, **kwargs: "Any") -> "AccessToken":  # pylint:disable=unused-argument
         """Asynchronously request an access token for `scopes`.
 
-        .. note:: This method is called by Azure SDK clients. It isn't intended for use in application code.
+        This method is called automatically by Azure SDK clients.
 
         :param str scopes: desired scopes for the access token. This method requires at least one scope.
         :rtype: :class:`azure.core.credentials.AccessToken`