+++ sdk/storage/azure-storage-blob/azure/storage/blob/_blob_service_client.py
@@ -383,6 +383,10 @@ class BlobServiceClient(StorageAccountHostsMixin):
         :param bool include_metadata:
             Specifies that container metadata to be returned in the response.
             The default value is `False`.
+        :keyword bool include_deleted:
+            Specifies that deleted containers to be returned in the response. This is for container restore enabled
+            account. The default value is `False`.
+            .. versionadded:: 12.4.0
         :keyword int results_per_page:
             The maximum number of container names to retrieve per API
             call. If the request does not specify the server will return up to 5,000 items.
@@ -401,6 +405,9 @@ class BlobServiceClient(StorageAccountHostsMixin):
                 :caption: Listing the containers in the blob service.
         """
         include = ['metadata'] if include_metadata else []
+        include_deleted = kwargs.pop('include_deleted', None)
+        if include_deleted:
+            include.append("deleted")
 
         timeout = kwargs.pop('timeout', None)
         results_per_page = kwargs.pop('results_per_page', None)
@@ -557,7 +564,7 @@ class BlobServiceClient(StorageAccountHostsMixin):
             **kwargs)
 
     @distributed_trace
-    def _undelete_container(self, deleted_container_name, deleted_container_version, new_name=None, **kwargs):
+    def undelete_container(self, deleted_container_name, deleted_container_version, **kwargs):
         # type: (str, str, str, **Any) -> ContainerClient
         """Restores soft-deleted container.
 
@@ -571,12 +578,14 @@ class BlobServiceClient(StorageAccountHostsMixin):
             Specifies the name of the deleted container to restore.
         :param str deleted_container_version:
             Specifies the version of the deleted container to restore.
-        :param str new_name:
+        :keyword str new_name:
             The new name for the deleted container to be restored to.
+            If not specified deleted_container_name will be used as the restored container name.
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
         :rtype: ~azure.storage.blob.ContainerClient
         """
+        new_name = kwargs.pop('new_name', None)
         container = self.get_container_client(new_name or deleted_container_name)
         try:
             container._client.container.restore(deleted_container_name=deleted_container_name,  # pylint: disable = protected-access