+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze_sentiment.py
@@ -690,7 +690,7 @@ class TestAnalyzeSentiment(TextAnalyticsTest):
     @GlobalTextAnalyticsAccountPreparer()
     @TextAnalyticsClientPreparer(client_kwargs={"api_version": TextAnalyticsApiVersion.V3_0})
     def test_opinion_mining_v3(self, client):
-        with pytest.raises(NotImplementedError) as excinfo:
+        with pytest.raises(ValueError) as excinfo:
             client.analyze_sentiment(["will fail"], show_opinion_mining=True)
 
         assert "'show_opinion_mining' is only available for API version v3.1-preview and up" in str(excinfo.value)