+++ sdk/eventgrid/azure-eventgrid/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Release History
 
-## 2.0.0b3 (Unreleased)
+## 2.0.0b3 (2020-10-05)
 
+  **Feature**
+  - Added support for Keyvault Event Types
+  - Added distributed tracing support for CloudEvents
 
 ## 2.0.0b2 (2020-09-24)